requirements.txt
streamlit
pandas
plotly
gspread
google-auth-oauthlib
google-cloud-firestore